+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change log
 
+## 1.1.0dev (unreleased)
+
+* `get-mathlib-cache` no longer understands `--rev`; if you want to use a
+  different mathlib version, edit your `leanproject.toml`.
+* Add `--fallback` to `get-cache` for traversing the git history to find an
+  approximate cache.
+* `get-cache` no longer modifies `lean` files in the working directory.
+* `mk-cache --force` no longer associates the cache with the current HEAD
+  revision, but instead creates a temporary commit in order to obtain a SHA1 id.
+
 ## 1.0.0 (2020-11-10)
 
 * Only look for .xz archives